refactor(code-block): decompose CodeBlockPlugin god class into focused collaborators
Split the 793-line CodeBlockPlugin into a thin (~205-line) orchestrator that
composes single-responsibility collaborators:

- TokenCache: per-block syntax-token caching + binary-search lookup + purge
- AutoPairController: bracket/quote pairing, overtype, wrap, dedent, and the
  mutable pair state (PairStack + deferred push/take queue)
- CodeBlockDecorations: focus + syntax-highlight decoration builder
- CodeBlockSchema: NodeSpec, input rule, toolbar, mark guard, paste, table-cell
- CodeBlockHighlighting: highlighter selection + language registry + service

decorations() is now a pure projection (cache purge moved to TokenCache.retain),
resolving the prior command-query separation smell. No public API change.

Adds focused unit tests for TokenCache and AutoPairController (incl. the
pre/post-transaction-space contract in syncOnStateChange).
